+++ src/Ydb.Sdk/CHANGELOG.md
@@ -3,7 +3,7 @@
 - ADO.NET: Added `MinPoolSize` setting to keep a minimum number of sessions ready in the PoolingSessionSource.
 - ADO.NET: Added `SessionIdleTimeout` to remove idle sessions from the PoolingSessionSource automatically.
 - ADO.NET: Made `PoolingSessionSource` faster and more reliable by using a lock-free LIFO stack.
-- ADO.NET: Added `BeginBulkUpsertImport` for batch upsert operations with transaction checks and integration tests.
+- ADO.NET: Added `BeginBulkUpsertImport` for batch upsert operations with transaction checks.
 - Optimization: On BadSession, do not invoke the `DeleteSession()` method.
 - Canceling AttachStream after calling the `DeleteSession` method.
 - Fixed bug: fixed issue where session was not deleted (`ClientTransportTimeout`).